Renaming package according to gemSpec name (i.e. "gemSpec_HSM-B")
